streamlined regression test data creation
