+++ src/prommis/conn/tests/test_connectivity.py
@@ -8,6 +8,7 @@ import pytest
 
 from prommis.conn import connectivity
 from prommis.uky.uky_flowsheet import build
+from prommis.conn.tests.connectivity_data import uky_csv, uky_mermaid
 
 
 def test_outputformats():
@@ -81,7 +82,7 @@ def test_ordering():
 
 
 @pytest.mark.unit
-def test_uky_csv(uky_csv, uky_mermaid):
+def test_uky_data(uky_csv, uky_mermaid):
     model = build()
     for fmt, ref_lines in (("csv", uky_csv), ("mermaid", uky_mermaid)):
         sio = StringIO()
@@ -90,134 +91,3 @@ def test_uky_csv(uky_csv, uky_mermaid):
         lines = data.split("\n")
         for line, ref_line in zip(lines, ref_lines):
             assert line == ref_line
-
-
-# Reference data
-
-
-@pytest.fixture
-def uky_csv():
-    return [
-        "Units,leach_mixer,leach,leach_liquid_feed,sl_sep1,leach_solid_feed,precipitator,sl_sep2,leach_sx_mixer,leach_filter_cake_liquid,leach_filter_cake,precip_sep,precip_purge,precip_sx_mixer,roaster,solex_cleaner_load,solex_cleaner_strip,cleaner_mixer,cleaner_org_make_up,acid_feed3,cleaner_sep,cleaner_purge,solex_rougher_load,load_sep,solex_rougher_scrub,rougher_mixer,rougher_org_make_up,acid_feed1,scrub_sep,solex_rougher_strip,acid_feed2,rougher_sep,sc_circuit_purge",
-        "Arcs,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,",
-        "leaching_feed_mixture,-1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "leaching_liq_feed,1,0,-1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "leaching_liquid_outlet,0,-1,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "leaching_sol_feed,0,1,0,0,-1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "leaching_solid_outlet,0,-1,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "precip_aq_outlet,0,0,0,0,0,-1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "precip_solid_outlet,0,0,0,0,0,-1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sl_sep1_liquid_outlet,0,0,0,-1,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sl_sep1_retained_liquid_outlet,0,0,0,-1,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sl_sep1_solid_outlet,0,0,0,-1,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sl_sep2_aq_purge,0,0,0,0,0,0,0,0,0,0,-1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sl_sep2_aq_recycle,0,0,0,0,0,0,0,0,0,0,-1,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sl_sep2_liquid_outlet,0,0,0,0,0,0,-1,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sl_sep2_retained_liquid_outlet,0,0,0,0,0,0,-1,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sl_sep2_solid_outlet,0,0,0,0,0,0,-1,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sx_cleaner_load_aq_feed,0,0,0,0,0,0,0,0,0,0,0,0,-1,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sx_cleaner_load_aq_outlet,0,0,0,0,0,0,0,1,0,0,0,0,0,0,-1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sx_cleaner_load_org_outlet,0,0,0,0,0,0,0,0,0,0,0,0,0,0,-1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sx_cleaner_mixed_org_recycle,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,-1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sx_cleaner_org_feed,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,-1,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sx_cleaner_strip_acid_feed,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,-1,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sx_cleaner_strip_aq_outlet,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,-1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sx_cleaner_strip_org_outlet,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,-1,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sx_cleaner_strip_org_purge,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,-1,1,0,0,0,0,0,0,0,0,0,0,0",
-        "sx_cleaner_strip_org_recycle,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,-1,0,0,0,0,0,0,0,0,0,0,0,0",
-        "sx_rougher_load_aq_feed,0,0,0,0,0,0,0,-1,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0",
-        "sx_rougher_load_aq_outlet,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,-1,1,0,0,0,0,0,0,0,0,0",
-        "sx_rougher_load_aq_recycle,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,-1,0,0,0,0,0,0,0,0,0",
-        "sx_rougher_load_org_outlet,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,-1,0,1,0,0,0,0,0,0,0,0",
-        "sx_rougher_mixed_org_recycle,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,-1,0,0,0,0,0,0,0",
-        "sx_rougher_org_feed,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,-1,0,0,0,0,0,0",
-        "sx_rougher_scrub_acid_feed,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,-1,0,0,0,0,0",
-        "sx_rougher_scrub_aq_outlet,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,-1,0,0,0,1,0,0,0,0",
-        "sx_rougher_scrub_aq_recycle,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,-1,0,0,0,0",
-        "sx_rougher_scrub_org_outlet,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,-1,0,0,0,0,1,0,0,0",
-        "sx_rougher_strip_acid_feed,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,-1,0,0",
-        "sx_rougher_strip_aq_outlet,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,-1,0,0,0",
-        "sx_rougher_strip_org_outlet,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,-1,0,1,0",
-        "sx_rougher_strip_org_purge,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,-1,1",
-        "sx_rougher_strip_org_recycle,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,-1,0",
-    ]
-
-
-@pytest.fixture
-def uky_mermaid():
-    return [
-        "flowchart LR",
-        "   Unit_B[leach_mixer]",
-        "   Unit_C[leach]",
-        "   Unit_D[leach_liquid_feed]",
-        "   Unit_E[sl_sep1]",
-        "   Unit_F[leach_solid_feed]",
-        "   Unit_G[precipitator]",
-        "   Unit_H[sl_sep2]",
-        "   Unit_I[leach_sx_mixer]",
-        "   Unit_J[leach_filter_cake_liquid]",
-        "   Unit_K[leach_filter_cake]",
-        "   Unit_L[precip_sep]",
-        "   Unit_M[precip_purge]",
-        "   Unit_N[precip_sx_mixer]",
-        "   Unit_O[roaster]",
-        "   Unit_P[solex_cleaner_load]",
-        "   Unit_Q[solex_cleaner_strip]",
-        "   Unit_R[cleaner_mixer]",
-        "   Unit_S[cleaner_org_make_up]",
-        "   Unit_T[acid_feed3]",
-        "   Unit_U[cleaner_sep]",
-        "   Unit_V[cleaner_purge]",
-        "   Unit_W[solex_rougher_load]",
-        "   Unit_X[load_sep]",
-        "   Unit_Y[solex_rougher_scrub]",
-        "   Unit_Z[rougher_mixer]",
-        "   Unit_AA[rougher_org_make_up]",
-        "   Unit_AB[acid_feed1]",
-        "   Unit_AC[scrub_sep]",
-        "   Unit_AD[solex_rougher_strip]",
-        "   Unit_AE[acid_feed2]",
-        "   Unit_AF[rougher_sep]",
-        "   Unit_AG[sc_circuit_purge]",
-        "   Unit_B --> Unit_C",
-        "   Unit_D --> Unit_B",
-        "   Unit_C --> Unit_E",
-        "   Unit_F --> Unit_C",
-        "   Unit_C --> Unit_E",
-        "   Unit_G --> Unit_H",
-        "   Unit_G --> Unit_H",
-        "   Unit_E --> Unit_I",
-        "   Unit_E --> Unit_J",
-        "   Unit_E --> Unit_K",
-        "   Unit_L --> Unit_M",
-        "   Unit_L --> Unit_N",
-        "   Unit_H --> Unit_L",
-        "   Unit_H --> Unit_O",
-        "   Unit_H --> Unit_O",
-        "   Unit_N --> Unit_P",
-        "   Unit_P --> Unit_I",
-        "   Unit_P --> Unit_Q",
-        "   Unit_R --> Unit_P",
-        "   Unit_S --> Unit_R",
-        "   Unit_T --> Unit_Q",
-        "   Unit_Q --> Unit_G",
-        "   Unit_Q --> Unit_U",
-        "   Unit_U --> Unit_V",
-        "   Unit_U --> Unit_R",
-        "   Unit_I --> Unit_W",
-        "   Unit_W --> Unit_X",
-        "   Unit_X --> Unit_B",
-        "   Unit_W --> Unit_Y",
-        "   Unit_Z --> Unit_W",
-        "   Unit_AA --> Unit_Z",
-        "   Unit_AB --> Unit_Y",
-        "   Unit_Y --> Unit_AC",
-        "   Unit_AC --> Unit_B",
-        "   Unit_Y --> Unit_AD",
-        "   Unit_AE --> Unit_AD",
-        "   Unit_AD --> Unit_N",
-        "   Unit_AD --> Unit_AF",
-        "   Unit_AF --> Unit_AG",
-        "   Unit_AF --> Unit_Z",
-        "",
-    ]
